split script into a package
