+++ vkmz/databases/bmrb-heavy.tsv
@@ -1,3 +1,4 @@
+mass	formula
 17.03465497	CH4
 28.02235974	C2H2
 29.01128877	CHN
