+++ vkmz/databases/bmrb-heavy_carbon.tsv
@@ -1,3 +1,4 @@
+mass	formula
 17.03465497	CH4
 28.01425388	CHN
 28.02235974	C2H2
