+++ vkmz/databases/bmrb-light.tsv
@@ -1,3 +1,4 @@
+mass	formula
 16.03130013	CH4
 26.01565006	C2H2
 27.01089904	CHN
